int64

+++ tests/test-finit.C
@@ -1,6 +1,8 @@
 #include "fflas-ffpack/utils/timer.h"
 #include "Matio.h"
 #include "fflas-ffpack/fflas/fflas.h"
+#include "fflas-ffpack/field/modular-int64.h"
+#include "fflas-ffpack/field/modular-balanced-int64.h"
 #include "fflas-ffpack/fflas-ffpack-config.h"
 #include "test-utils.h"
 #include "assert.h"
@@ -96,7 +98,7 @@ int main(int ac, char **av) {
 	static size_t m = 300 ;
 	static size_t n = 301 ;
 	static size_t k = 300 ;
-	static size_t p = 7;
+	static unsigned long p = 7;
 	int seed = (int) time(NULL);
 
 	static Argument as[] = {
